requirements.txt
astroid==1.4.6
colorama==0.3.7
coverage==4.1
Django==1.9.7
django-cors-headers==1.1.0
django-filter==0.13.0
djangorestframework==3.3.3
drfdocs==0.0.11
lazy-object-proxy==1.2.2
mongoengine==0.10.6
pylint==1.6.4
pymongo==3.2.2
six==1.10.0
wrapt==1.10.8
newrelic==2.66.0.49
PyJWT==1.4.0
Fabric3==1.11.1.post1
gunicorn==19.6.0
pymssql==2.1.3
redis==2.10.5
SQLAlchemy==1.0.14
requests==2.10.0

